requirements.txt
anthropic
datasets
google-generativeai
matplotlib
openai
pandas
torch
transformers
seaborn
sentence-transformers
